# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "eKYC API",
    "description": "An API for eKYC prototyping",
    "termsOfService": "https://aws.amazon.com/asl/",
    "contact": {
      "name": "Amazon Web Services",
      "url": "https://aws.amazon.com",
      "email": "opensource-codeofconduct@amazon.com"
    },
    "license": {
      "name": "Use under Amazon Software License",
      "url": "https://aws.amazon.com/asl/"
    },
    "version": "v2"
  },
  "paths": {
...
    "/api/session/url": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "Gets a presigned URL to allow uploads using the HTTP PUT verb.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "The session ID that is being used across multiple requests.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s3Key",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/compare": {
      "post": {
        "tags": [
          "Session"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompareDocumentWithSelfie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareDocumentWithSelfie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareDocumentWithSelfie"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/selfie": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Submit a selfie image for liveness checking.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "The session ID that is being used across multiple requests.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s3Key",
            "in": "query",
            "description": "The key of the selfie stored in S3. Must be a valid image of a person with eyes open.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/session/document": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Submits a document for liveness verification.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "The session ID that is being used across multiple requests.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s3Key",
            "in": "query",
            "description": "The s3 key where the document has been uploaded.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expectedDocumentType",
            "in": "query",
            "description": "The type of document that has been uploaded.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/session/nosepoint": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Submits a nose pointing image for liveness verification.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "The session ID that is being used across multiple requests.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s3Key",
            "in": "query",
            "description": "The key of the image stored in S3.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/session/eyesclosed": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Submits the eyes closed image for liveness verification.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "The session ID that is being used across multiple requests.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s3Key",
            "in": "query",
            "description": "The key of the image stored in S3.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompareDocumentWithSelfie": {
        "type": "object",
        "properties": {
          "isSimilar": {
            "type": "boolean",
            "nullable": true
          },
          "similarity": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}